Remove code-toolbar by using a simplified markdown renderer in settings
